requirements.txt
requests==2.31.0
python-dotenv==0.20.0
mutagen==1.47.0